backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
websockets
requests
pandas
numpy
openai
python-jose
apscheduler